Refactor router; move project middleware
