+++ router.js
@@ -1,4 +1,4 @@
 const router = new require('koa-router')()
-router.post('/project', require('./project'))
+router.post('/project', require('./commands/project'))
 
 module.exports = router
